backend/pyproject.toml
[tool.poetry]
name = "vtt-backend"
version = "0.1.0"
description = "Voice to Tool - Backend service for voice processing and tool integration"
authors = ["Marcel Lewi <marchulew@gmail.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
langchain = "0.1.12"
langchain-openai = "0.0.5"
langgraph = "0.0.26"
langchain-core = "0.1.31"
openai-whisper = "20240930"
sqlmodel = "^0.0.24"
alembic = "^1.15.2"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 